whatsapp-demo/requirements.txt
# WhatsApp AI Demo — Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
